use PropertyFloat instead of PropertyInt

+++ Source/ACE.Server/WorldObjects/Player_Database.cs
@@ -44,7 +44,7 @@ namespace ACE.Server.WorldObjects
 
         private void SetPropertiesAtLogOut()
         {
-            LogoffTimestamp = (int)Time.GetUnixTime();
+            LogoffTimestamp = Time.GetUnixTime();
             // These properties are used with offline players to determine passup rates
             SetProperty(PropertyInt.CurrentLoyaltyAtLastLogoff, (int)GetCreatureSkill(Skill.Loyalty).Current);
             SetProperty(PropertyInt.CurrentLeadershipAtLastLogoff, (int)GetCreatureSkill(Skill.Leadership).Current);
